Implemented flush-on-newline, paragraph-attribute.
Implemented “flush-output-on-newline” configuration option. This will
ensure that any output is flushed to history and interface before a
newline. This allows interfaces to store paragraph-relevant attribtues
in the history with minor fuzz.

Added new metadata-type “paragraph attribute” for the purpose
described above.

Minor changes in processing when converting font 3 to unicode.