Fixed print_table for lower window / Sherlock.
This change tries to merge the new Z-Spec definition for print_table
and to make "Sherlock" work with it. According to the new spec,
print_table shouldn't be used in the lower window. Sherlock however
appears to require this to display error messages correctly. libfizmo
now allows the print_table opcode to work in the lower window, but
simply shows the opcode's text without any cursor movement.